HACK: Miyoo Mini Plus Uboot is special
The Miyoo Mini Plus stock uboot can only properly load an Image that is wrapped in a uImage. This means that the
zImage appended method, and the kernel.fit style of loading do not work well with the stock uboot.

Since the stock uboot is the only one that can get a screen working, we have to adopt this workaround until we can replace
the stock uboot with our own.

One way that you can load a regular zImage is by using the 'go' command in the uboot instead of the 'bootm' command. However
this proves to be very unreliable and frustrating